add keytrack panning
The samples are mono, but I wanted to create a better sense of space, so I implemented stereo keytrack panning with low notes on the left and high notes on the right of the stereo field. The amount of keytrack panning can be controlled using CC#21 and is set to 50% (64) by default.